Fix #79 and some bugs
fix the bug causing the width of xterm-screen to decrease slightly when dragging the resize-box vertically.
fix Single window resize fires many resize events #79. The width of resize-box is adjusted to match that of the flex parent element when the parent is smaller than resize-box now.
bump up the version of xterm addons.